Add placeholder for "official" tutorials
The ideas is that these would be produced/validated
by someone we know, and expected to be maintained.

I expect we would host them here, so there's some
change/quality control.

Someone with energy might even script each tutorial
and create a CI testcase that checks the output...

Hopefully these can supplement and/or replace the
random links I've put on /software.