Add helper script used to run the debug webserver
SERVER=hostname debug_webserver

Allows a browser on a desktop to access the server; the
server's default is localhost, which reduces one to a
browser on the server.  Browser performance can be quite
bad compared to an external browser in that case.

BUT:  Don't expose the server to the public internet!!