Re-organize software and simulators.
There's now a page listing all the simulators (that I could find easily).
It has whatever simulator user doc was is in the simulator repo.

When know of software for a family, there's a page for it that
applicable simulators link to.
E.g. all the PDP-10 simulators point to a single page for TOPS-10,
TOPS-20, and ITS.

Hopefully, this will work.